More robust root finding algorithm
fsolve is unbound, leading to invalid values of theta in the
determination of the Jacobian. The brentq method finds the
zero-crossing in a known interval. As K_r is underestimated at
theta=theta_r and is overestimated at theta=theta_s there
must be a zero-crossing of the error function in between.